frontend/requirements.txt
streamlit==1.37.1
requests==2.32.3
python-dotenv==1.0.1
langchain-core>=0.3.0,<0.4.0